requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
frozenlist==1.6.0
geoip2==5.1.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
maxminddb==2.7.0
multidict==6.4.4
openpyxl==3.1.5
packaging==25.0
priority==2.0.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
requests==2.32.3
service-identity==24.2.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
waitress==3.0.2
whitenoise==6.9.0
wsproto==1.2.0
yarl==1.20.0
zope.interface==7.2
